+++ src/providers/semrush.js
@@ -1,0 +1,16 @@
+/**
+ * semrush.js
+ */
+
+module.exports = {
+  name: 'SemrushBot',
+  testAddresses: ['85.208.98.53'],
+  ipv4: {
+    addresses: [],
+    ranges: ['85.208.98.32/28', '85.208.98.48/28'],
+  },
+  ipv6: {
+    addresses: [],
+    ranges: [],
+  },
+};